+++ start.sh
@@ -8,10 +8,10 @@ then
 fi
 
 mkdir -p ./dev
-TARANTOOL_ALIAS=srv-1 TARANTOOL_WORKDIR=dev/3301 TARANTOOL_ADVERTISE_URI=localhost:3301 TARANTOOL_HTTP_PORT=8081 $SCRIPT & echo $! >> ./dev/pids
-TARANTOOL_ALIAS=srv-2 TARANTOOL_WORKDIR=dev/3302 TARANTOOL_ADVERTISE_URI=localhost:3302 TARANTOOL_HTTP_PORT=8082 $SCRIPT & echo $! >> ./dev/pids
-TARANTOOL_ALIAS=srv-3 TARANTOOL_WORKDIR=dev/3303 TARANTOOL_ADVERTISE_URI=localhost:3303 TARANTOOL_HTTP_PORT=8083 $SCRIPT & echo $! >> ./dev/pids
-TARANTOOL_ALIAS=srv-4 TARANTOOL_WORKDIR=dev/3304 TARANTOOL_ADVERTISE_URI=localhost:3304 TARANTOOL_HTTP_PORT=8084 $SCRIPT & echo $! >> ./dev/pids
-TARANTOOL_ALIAS=srv-5 TARANTOOL_WORKDIR=dev/3305 TARANTOOL_ADVERTISE_URI=localhost:3305 TARANTOOL_HTTP_PORT=8085 $SCRIPT & echo $! >> ./dev/pids
+$SCRIPT --alias srv-1 --workdir dev/3301 --advertise-uri localhost:3301 --http-port 8081 & echo $! >> ./dev/pids
+$SCRIPT --alias srv-2 --workdir dev/3302 --advertise-uri localhost:3302 --http-port 8082 & echo $! >> ./dev/pids
+$SCRIPT --alias srv-3 --workdir dev/3303 --advertise-uri localhost:3303 --http-port 8083 & echo $! >> ./dev/pids
+$SCRIPT --alias srv-4 --workdir dev/3304 --advertise-uri localhost:3304 --http-port 8084 & echo $! >> ./dev/pids
+$SCRIPT --alias srv-5 --workdir dev/3305 --advertise-uri localhost:3305 --http-port 8085 & echo $! >> ./dev/pids
 sleep 1.5
 echo "All instances started!"