+++ test/integration/srv_multisharding.lua
@@ -19,17 +19,12 @@ errors.set_deprecation_handler(function(err)
 end)
 
 local ok, err = cluster.cfg({
-    alias = os.getenv('TARANTOOL_ALIAS'),
-    workdir = os.getenv('TARANTOOL_WORKDIR'),
-    advertise_uri = os.getenv('TARANTOOL_ADVERTISE_URI') or 'localhost:3301',
-    cluster_cookie = os.getenv('TARANTOOL_CLUSTER_COOKIE'),
     bucket_count = nil,
     vshard_groups = {
         -- both notations are valid
         ['cold'] = {bucket_count = 2000},
         'hot',
     },
-    http_port = os.getenv('TARANTOOL_HTTP_PORT') or 8081,
     roles = {
         'cluster.roles.vshard-storage',
         'cluster.roles.vshard-router',